SBH/BACKEND/MyFunctionApp/requirements.txt
python-dotenv
pandas
gmft
google-generativeai